maint: fix -Wformat-exposed errors in uses of "panic"
* sed/utils.h (panic): Add __printf__ attribute, so that compilers
can diagnose format/arg-type errors.  This exposed the following:
* sed/utils.c (ck_fwrite): Don't try to print size_t via %u.
Instead, use %llu and a cast to unsigned long long.
* sed/mbcs.c (is_mb_char): Cast an "int" to unsigned int, to avoid
mismatch with %x.

+++ sed/mbcs.c
@@ -59,7 +59,8 @@ is_mb_char (int ch, mbstate_t *cur_stat)
       return 1;
 
     default: /* Should never happen, as per mbrtowc(3) documentation */
-      panic ("is_mb_char: mbrtowc (0x%x) returned %d",ch,result);
+      panic ("is_mb_char: mbrtowc (0x%x) returned %d",
+             (unsigned int) ch, result);
     }
 }
